requirements.txt
streamlit>=1.30.0
apify-client>=1.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
requests>=2.31.0
yt-dlp>=2024.1.0
